requirements.txt
PyQt5
Pillow
matplotlib
mutagen
pyinstaller